print out the value of pi, as float ready but give back 3.00 instead of 3.14 problem is fixed with the atof function

+++ week-03/day-2/week3_day2_e04/main.c
@@ -9,7 +9,7 @@ int main()
 
 	strcpy(str, "3.14");
 
-    val = atoi(str);
+    val = atof(str);
 	printf("String value: %s , float value: %.2f", str , val);
 
 	//TODO: print out the value of pi, first as a string, then a float value.
